requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==3.3.6
async-timeout==5.0.1
attrs==24.3.0
billiard==4.2.1
boto==2.49.0
cache==1.0.3
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect2==0.7.0
dataclass-wizard==0.22.3
diff-match-patch==20241021
dill==0.3.9
Django==5.1.3
django-allauth==65.3.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-debug-toolbar==5.0.1
django-filter==24.3
django-import-export==4.3.3
django-pdf==1.0.2
django-redis==5.4.0
django-search==1.0
django-simple-history==3.7.0
django-timezone-field==7.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
easy-colors==0.0.1
et_xmlfile==2.0.0
ghettoq==0.4.5
html5lib==1.1
idna==3.10
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lxml==5.3.0
mailhog==0.1.1
mccabe==0.7.0
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pylint==3.3.2
pypdf==5.1.0
python-bidi==0.6.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
redis==5.2.1
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
send-mail==0.0.2
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.1
svglib==1.5.1
tablib==3.7.0
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.16
